# Improve performance of immutable Map.contains