Map.contains defaults to ``` get(key).isDefined ``` generating an extra Option wrapper on each use (which is immediately discarded).    Override the base implementation to reduce memory churn.

Performance benchmark.
```
before 
[info] Benchmark              (size)  Mode  Cnt      Score      Error  Units
[info] MapBenchmark.contains       1  avgt   20      2.610 ±    0.138  ns/op
[info] MapBenchmark.contains      10  avgt   20     30.473 ±    1.233  ns/op
[info] MapBenchmark.contains     100  avgt   20   1549.097 ±   65.146  ns/op
[info] MapBenchmark.contains    1000  avgt   20  21638.244 ± 1021.994  ns/op

after
[info] Benchmark              (size)  Mode  Cnt      Score     Error  Units
[info] MapBenchmark.contains       1  avgt   20      2.682 ±   0.116  ns/op
[info] MapBenchmark.contains      10  avgt   20     31.703 ±   1.563  ns/op
[info] MapBenchmark.contains     100  avgt   20   1068.583 ±  53.095  ns/op
[info] MapBenchmark.contains    1000  avgt   20  14310.988 ± 758.181  ns/op
```