@retronym 2017-02-02
> Great minds think alike @retronym?

Either that, or simple minds seldom differ... 😉 

We'll need the whitelist changes from my PR, so if you don't mind, I'll close this PR and we can continue the review and compatibilty discussions over there.